+++ src/lib/CoinbasePro/Types.hs
@@ -12,7 +12,7 @@ module CoinbasePro.Types
     , Size (..)
     , Volume (..)
     , TradeId (..)
-    , Funds
+    , Funds (..)
     , OrderType (..)
     , CreatedAt (..)
     , Candle (..)