+++ changelog.md
@@ -1,3 +1,10 @@
+# [Unreleased]
+ - Added `placeLimitOrder` and `placeMarketOrder`, allowing for orders to
+  optionally specify stop and stop price. `placeLimitOrder` exposes
+  'cancelAfter' option for limit orders and `placeMarketOrders` allows for
+  market orders with defined funds but no size.
+ - `placeOrder` remains for backwards-compatibility but is now deprecated.
+
 # Version 0.7.2.0
  - Upgraded to stack lts-14.17